pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hyperliquid-copybot"
version = "0.1.0"
description = "Production-ready Hyperliquid copy trading bot"
requires-python = ">=3.9"
dependencies = [
    "hyperliquid-python-sdk>=0.4.0",
    "websockets>=12.0",
    "aiohttp>=3.9",
    "structlog>=24.0",
    "pyyaml>=6.0",
    "aiosqlite>=0.19",
    "python-dotenv>=1.0",
    "boto3>=1.34",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "ruff>=0.3",
]

[project.scripts]
copybot = "copybot.main:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
